Remove plugin packages
