--- lyra-plugin/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "lyra-plugin"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-lyra-common = { path = "../lyra-common" }
-anyhow = {workspace = true}
-egui = { workspace = true }
-nucleo-matcher = {workspace = true}
-parking_lot = {workspace = true}
-serde_json = {workspace = true}
-serde = {workspace = true}
-toml = {workspace = true}
-tracing = {workspace = true}
-
-# Calc
-arboard = "3.3.0"
-calc = { version = "0.4.0", default-features = false }
-lalrpop-util = { version = "0.19.12", features = ["lexer"] }
-
-# Apps
-base64 = {workspace = true}
-glob = "0.3.1"
-icns = "0.3.1"
-plist = "1.6.0"
-
-# Apps / Webq
-open = {workspace=true}
