+++ lyra-ui/Cargo.toml
@@ -11,7 +11,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-lyra-plugin = { path = "../lyra-plugin" }
+lyra-common = { path = "../lyra-common" }
 anyhow = { workspace = true }
 base64 = { workspace = true }
 chrono = "0.4.31"
@@ -33,6 +33,19 @@ tracing-appender = "0.2.3"
 tracing-subscriber = "0.3.18"
 tray-icon = "0.11.2"
 winit = "0.29.10"
+
+# Calc
+arboard = "3.3.0"
+calc = { version = "0.4.0", default-features = false }
+lalrpop-util = { version = "0.19.12", features = ["lexer"] }
+
+# Apps
+glob = "0.3.1"
+icns = "0.3.1"
+plist = "1.6.0"
+
+# Apps / Webq
+open = {workspace=true}
 
 [target."cfg(target_os = \"linux\")".dependencies]
 gtk = "0.18"
